Update cwe_wartime_measures_button.txt war bonds
War bonds had a bug where it stopped having any effects when the investment pool was greater than 10 billion, due to "		if = {
			limit = { 
				investment_pool > 1000000000
				investment_pool <= 10000000000"
This is the highest bracket and it has a upper limit, other possible solutions would be to remove the upper bracket but then the transfer amount would be limited to 100 million regardless to how large the investment pool becomes. Another solution would be to simply add more if statments with higher numbers.